refactor: move core folder component code to index
